+++ tensorlayer/layers/convolution/quan_conv_bn.py
@@ -16,12 +16,12 @@ from tensorlayer import logging
 
 from tensorlayer.decorators import deprecated_alias
 
-__all__ = ['QuanConv2dWithBN']
+__all__ = ['QuantizedConv2dWithBN']
 
 
-class QuanConv2dWithBN(Layer):
-    """The :class:`QuanConv2dWithBN` class is a quantized convolutional layer with BN, which weights are 'bitW' bits and the output of the previous layer
-    are 'bitA' bits while inferencing.
+class QuantizedConv2dWithBN(Layer):
+    """The :class:`QuantizedConv2dWithBN` class is a quantized convolutional layer with BN, which weights are 'bitW' bits and
+    the output of the previous layer are 'bitA' bits while inferencing.
 
     Note that, the bias vector would keep the same.
 
@@ -87,10 +87,10 @@ class QuanConv2dWithBN(Layer):
     >>> import tensorlayer as tl
     >>> x = tf.placeholder(tf.float32, [None, 256, 256, 3])
     >>> net = tl.layers.InputLayer(x, name='input')
-    >>> net = tl.layers.QuanConv2dWithBN(net, 64, (5, 5), (1, 1),  act=tf.nn.relu, padding='SAME', is_train=is_train, bitW=bitW, bitA=bitA, name='qcnnbn1')
+    >>> net = tl.layers.QuantizedConv2dWithBN(net, 64, (5, 5), (1, 1),  act=tf.nn.relu, padding='SAME', is_train=is_train, bitW=bitW, bitA=bitA, name='qcnnbn1')
     >>> net = tl.layers.MaxPool2d(net, (3, 3), (2, 2), padding='SAME', name='pool1')
     ...
-    >>> net = tl.layers.QuanConv2dWithBN(net, 64, (5, 5), (1, 1), padding='SAME', act=tf.nn.relu, is_train=is_train,  bitW=bitW, bitA=bitA, name='qcnnbn2')
+    >>> net = tl.layers.QuantizedConv2dWithBN(net, 64, (5, 5), (1, 1), padding='SAME', act=tf.nn.relu, is_train=is_train,  bitW=bitW, bitA=bitA, name='qcnnbn2')
     >>> net = tl.layers.MaxPool2d(net, (3, 3), (2, 2), padding='SAME', name='pool2')
     ...
     """
@@ -114,14 +114,14 @@ class QuanConv2dWithBN(Layer):
             use_gemm=False,
             W_init=tf.truncated_normal_initializer(stddev=0.02),
             W_init_args=None,
-            use_cudnn_on_gpu=None,
+            use_cudnn_on_gpu=True,
             data_format=None,
             name='quan_cnn2d_bn',
     ):
-        super(QuanConv2dWithBN, self).__init__(prev_layer=prev_layer, act=act, W_init_args=W_init_args, name=name)
+        super(QuantizedConv2dWithBN, self).__init__(prev_layer=prev_layer, act=act, W_init_args=W_init_args, name=name)
 
         logging.info(
-            "QuanConv2dWithBN %s: n_filter: %d filter_size: %s strides: %s pad: %s act: %s " % (
+            "QuantizedConv2dWithBN %s: n_filter: %d filter_size: %s strides: %s pad: %s act: %s " % (
                 self.name, n_filter, filter_size, str(strides), padding, self.act.__name__
                 if self.act is not None else 'No Activation'
             )
@@ -137,12 +137,12 @@ class QuanConv2dWithBN(Layer):
             raise ValueError("len(strides) should be 2.")
 
         try:
-            pre_channel = int(prev_layer.outputs.get_shape()[-1])
-        except Exception:  # if pre_channel is ?, it happens when using Spatial Transformer Net
-            pre_channel = 1
+            input_channels = int(prev_layer.outputs.get_shape()[-1])
+        except TypeError:  # if input_channels is ?, it happens when using Spatial Transformer Net
+            input_channels = 1
             logging.warning("[warnings] unknow input channels, set to 1")
 
-        shape = (filter_size[0], filter_size[1], pre_channel, n_filter)
+        shape = (filter_size[0], filter_size[1], input_channels, n_filter)
         strides = (1, strides[0], strides[1], 1)
 
         with tf.variable_scope(name):
@@ -219,8 +219,7 @@ class QuanConv2dWithBN(Layer):
             self.outputs = self._apply_activation(self.outputs)
 
         self._add_layers(self.outputs)
-
-        self._add_params([W, scale_para, offset_para, moving_mean, moving_variance])
+        self._add_params(self._local_weights)
 
 
 def w_fold(w, gama, var, epsilon):