+++ tests/test_layers_normalization.py
@@ -20,9 +20,9 @@ def model(x, is_train=True, reuse=False):
         n = tl.layers.LayerNormLayer(n, reuse=reuse, name='norm_layer')
         n = tl.layers.InstanceNormLayer(n, name='norm_instance')
         n = tl.layers.SwitchNormLayer(n, name='switchnorm')
-        n = tl.layers.QuanConv2dWithBN(n, n_filter=3, is_train=is_train, name='quan_cnn_with_bn')
+        n = tl.layers.QuantizedConv2dWithBN(n, n_filter=3, is_train=is_train, name='quan_cnn_with_bn')
         n = tl.layers.FlattenLayer(n, name='flatten')
-        n = tl.layers.QuanDenseLayerWithBN(n, n_units=10, name='quan_dense_with_bn')
+        n = tl.layers.QuantizedDenseWithBN(n, n_units=10, name='quan_dense_with_bn')
     return n